pyproject.toml
[project]
name = "writers-ensemble"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.49.0",
    "jinja2>=3.1.6",
]
